tests/wpt/web-platform-tests/speech-api/historical.html

<!doctype html>
<title>Historical Speech API features</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
[
  "SpeechRecognitionError",
  "webkitSpeechGrammar",
  "webkitSpeechGrammarList",
  "webkitSpeechRecognition",
  "webkitSpeechRecognitionError",
  "webkitSpeechRecognitionEvent",
].forEach(name => {
  test(function() {
    assert_false(name in window);
  }, name + " interface should not exist");
});
</script>
